+++ android/gltfio-android/src/main/java/com/google/android/filament/gltfio/ResourceLoader.java
@@ -45,7 +45,7 @@ public class ResourceLoader {
      */
     public ResourceLoader(@NonNull Engine engine) {
         long nativeEngine = engine.getNativeObject();
-        mNativeObject = nCreateResourceLoader(nativeEngine, false, false);
+        mNativeObject = nCreateResourceLoader(nativeEngine, false, false, false);
     }
 
     /**
@@ -54,14 +54,15 @@ public class ResourceLoader {
      * @param engine the engine that gets passed to all builder methods
      * @param normalizeSkinningWeights scale non-conformant skinning weights so they sum to 1
      * @param recomputeBoundingBoxes use computed bounding boxes rather than the ones in the asset
+     * @param ignoreBindTransform ignore skinned primitives bind transform when compute bounding boxes
      * @throws IllegalAccessException
      * @throws InvocationTargetException
      */
     public ResourceLoader(@NonNull Engine engine, boolean normalizeSkinningWeights,
-            boolean recomputeBoundingBoxes) {
+            boolean recomputeBoundingBoxes, boolean ignoreBindTransform) {
         long nativeEngine = engine.getNativeObject();
         mNativeObject = nCreateResourceLoader(nativeEngine, normalizeSkinningWeights,
-                recomputeBoundingBoxes);
+                recomputeBoundingBoxes, ignoreBindTransform);
     }
 
     /**
@@ -166,7 +167,7 @@ public class ResourceLoader {
     }
 
     private static native long nCreateResourceLoader(long nativeEngine,
-            boolean normalizeSkinningWeights, boolean recomputeBoundingBoxes);
+            boolean normalizeSkinningWeights, boolean recomputeBoundingBoxes, boolean ignoreBindTransform);
     private static native void nDestroyResourceLoader(long nativeLoader);
     private static native void nAddResourceData(long nativeLoader, String url, Buffer buffer,
             int remaining);