+++ libs/filabridge/include/private/filament/Variant.h
@@ -185,6 +185,10 @@ namespace filament {
             if (isValidDepthVariant(variantKey)) {
                 return variantKey;
             }
+            // if shadow receiver is disabled, turn off VSM
+            if (isLit && !(variantKey & SHADOW_RECEIVER)) {
+                return variantKey & ~VSM;
+            }
             // when the shading mode is unlit, remove all the lighting variants
             return isLit ? variantKey : (variantKey & UNLIT_MASK);
         }