+++ libs/gltfio/src/MorphHelper.cpp
@@ -16,11 +16,12 @@
 
 #include "MorphHelper.h"
 
-#include <filament/BufferObject.h>
 #include <filament/RenderableManager.h>
 #include <filament/VertexBuffer.h>
+#include <utils/JobSystem.h>
+#include <math/norm.h>
+#include <filament/MorphTargetBuffer.h>
 
-#include "GltfEnums.h"
 #include "TangentsJob.h"
 
 using namespace filament;
@@ -45,14 +46,6 @@ static bool isSubsetOf(ubyte4 a, ubyte4 b) {
     return true;
 }
 
-static int indexOf(int a, ubyte4 b) {
-    if (a == b.x) return 0;
-    if (a == b.y) return 1;
-    if (a == b.z) return 2;
-    if (a == b.w) return 3;
-    return -1;
-}
-
 MorphHelper::MorphHelper(FFilamentAsset* asset, FFilamentInstance* inst) : mAsset(asset),
         mInstance(inst) {
     NodeMap& sourceNodes = asset->isInstanced() ? asset->mInstances[0]->nodeMap : asset->mNodeMap;
@@ -62,7 +55,7 @@ MorphHelper::MorphHelper(FFilamentAsset* asset, FFilamentInstance* inst) : mAsse
         if (mesh) {
             cgltf_primitive const* prims = mesh->primitives;
             for (cgltf_size pi = 0, count = mesh->primitives_count; pi < count; ++pi) {
-                addPrimitive(mesh, pi, &mMorphTable[pair.second]);
+                addPrimitive(mesh, pi, pair.second);
             }
         }
     }
@@ -72,96 +65,50 @@ MorphHelper::~MorphHelper() {
     auto engine = mAsset->mEngine;
     for (auto& entry : mMorphTable) {
         for (auto& prim : entry.second.primitives) {
-            for (auto& target : prim.targets) {
-                engine->destroy(target.bufferObject);
+            if (prim.targets) {
+                engine->destroy(prim.targets);
             }
         }
     }
 }
 
-void MorphHelper::applyWeights(Entity entity, float const* weights, size_t count) noexcept {
+void MorphHelper::setWeights(Entity entity, float const* weights, int count) noexcept {
     auto& engine = *mAsset->mEngine;
-    auto renderableManager = &engine.getRenderableManager();
-    auto renderable = renderableManager->getInstance(entity);
+    auto& rcm = engine.getRenderableManager();
+    rcm.setMorphWeights(rcm.getInstance(entity), weights, count);
+}
 
-    // If there are 4 or fewer targets, we can simply re-use the original VertexBuffer.
-    if (count <= 4) {
-        float4 vec{};
-        for (size_t i = 0; i < count; i++) {
-            vec[i] = weights[i];
-        }
-        renderableManager->setMorphWeights(renderable, vec);
-        return;
-    }
-
-    // We honor up to 255 weights because our set representation is a 4-tuple of bytes, with one
-    // slot reserved for a sentinel value. Note that 255 is much more than the glTF min spec of 4.
-    count = std::min(count, size_t(255));
-
-    // Make a copy of the weights because we want to re-order them.
-    auto& sorted = mPartiallySortedWeights;
-    sorted.clear();
-    sorted.insert(sorted.begin(), weights, weights + count);
-
-    // Find the four highest weights in O(n) by doing a partial sort.
-    std::nth_element(sorted.begin(), sorted.begin() + 4, sorted.end(), [](float a, float b) {
-        return a > b;
-    });
-
-    // Find the "primary indices" which are the indices of the four highest weights. This is O(n).
-    ubyte4 primaryIndices = {kUnused, kUnused, kUnused, kUnused};
-    for (size_t index = 0, primary = 0; index < count && primary < 4; ++index) {
-        const float w = weights[index];
-        if (w > 0 && (w == sorted[0] || w == sorted[1] || w == sorted[2] || w == sorted[3])) {
-            primaryIndices[primary++] = index;
+int MorphHelper::getTargetCount(Entity entity) const noexcept {
+    if (mMorphTable.count(entity)) {
+        auto& primitive = mMorphTable.at(entity).primitives;
+        if (!primitive.empty() && primitive[0].targets) {
+            return primitive[0].targets->getCount();
         }
     }
-
-    // Swap out the buffer objects for the primary indices.
-    for (const auto& prim : mMorphTable[entity].primitives) {
-        for (const auto& target : prim.targets) {
-            const int index = indexOf(target.morphTargetIndex, primaryIndices);
-            if (index > -1) {
-                assert_invariant(primaryIndices[index] == target.morphTargetIndex);
-                VertexBuffer* vb = prim.vertexBuffer;
-                const int bufferObjectSlot = target.type == cgltf_attribute_type_position ?
-                        prim.positions[index] : prim.tangents[index];
-
-                // Slot 0 is always used for the base position so if it's getting clobbered, then
-                // something is seriously wrong. Should never occur, assert just in case.
-                assert_invariant(bufferObjectSlot != 0);
-
-                vb->setBufferObjectAt(engine, bufferObjectSlot, target.bufferObject);
-
-                // Do not break out early because there could be more than one target entry for this
-                // particular target index (e.g. positions + tangents).
-            }
-        }
-    }
-
-    // Finally, set the 4-tuple uniform for the weight values by derefing the primary indices.
-    // Note that we first create a "safe set" by replacing the unused sentinel with zero.
-    float4 highest;
-    for (int i = 0; i < 4; i++) {
-        highest[i] = (primaryIndices[i] == kUnused) ? 0 : weights[primaryIndices[i]];
-    }
-    renderableManager->setMorphWeights(renderable, highest);
+    return 0;
 }
 
 // This method copies various morphing-related data from the FilamentAsset MeshCache primitive
 // (which lives in transient memory) into the MorphHelper primitive (which will stay resident).
-void MorphHelper::addPrimitive(cgltf_mesh const* mesh, int primitiveIndex, TableEntry* entry) {
+void MorphHelper::addPrimitive(cgltf_mesh const* mesh, int primitiveIndex, Entity entity){
+    auto& entry = mMorphTable[entity];
     auto& engine = *mAsset->mEngine;
     const cgltf_primitive& prim = mesh->primitives[primitiveIndex];
     const auto& gltfioPrim = mAsset->mMeshCache.at(mesh)[primitiveIndex];
     VertexBuffer* vertexBuffer = gltfioPrim.vertices;
+    entry.primitives.push_back({ nullptr });
 
-    entry->primitives.push_back({ vertexBuffer });
-    auto& morphHelperPrim = entry->primitives.back();
+    auto& morphHelperPrim = entry.primitives.back();
 
-    for (int i = 0; i < 4; i++) {
-        morphHelperPrim.positions[i] = gltfioPrim.morphPositions[i];
-        morphHelperPrim.tangents[i] = gltfioPrim.morphTangents[i];
+    if (prim.targets_count) {
+        morphHelperPrim.targets = MorphTargetBuffer::Builder()
+                .vertexCount(vertexBuffer->getVertexCount())
+                .count(prim.targets_count)
+                .build(engine);
+
+        auto& rcm = engine.getRenderableManager();
+        rcm.setMorphTargetBufferAt(rcm.getInstance(entity),
+                primitiveIndex, morphHelperPrim.targets);
     }
 
     const cgltf_accessor* previous = nullptr;
@@ -175,19 +122,15 @@ void MorphHelper::addPrimitive(cgltf_mesh const* mesh, int primitiveIndex, Table
                 continue;
             }
             if (atype == cgltf_attribute_type_normal) {
-
                 // TODO: use JobSystem for this, like what we do for non-morph tangents.
                 TangentsJob job;
                 TangentsJob::Params params = { .in = { &prim, targetIndex } };
                 TangentsJob::run(&params);
 
                 if (params.out.results) {
-                    const size_t size = params.out.vertexCount * sizeof(short4);
-                    BufferObject* bufferObject = BufferObject::Builder().size(size).build(engine);
-                    VertexBuffer::BufferDescriptor bd(params.out.results, size, FREE_CALLBACK);
-                    bufferObject->setBuffer(engine, std::move(bd));
-                    params.out.results = nullptr;
-                    morphHelperPrim.targets.push_back({bufferObject, targetIndex, atype});
+                    morphHelperPrim.targets->setTangentsAt(engine, targetIndex,
+                            params.out.results, params.out.vertexCount);
+                    free(params.out.results);
                 }
                 continue;
             }
@@ -201,19 +144,16 @@ void MorphHelper::addPrimitive(cgltf_mesh const* mesh, int primitiveIndex, Table
                 if (accessor->buffer_view) {
                     auto bufferData = (const uint8_t*) accessor->buffer_view->buffer->data;
                     assert_invariant(bufferData);
-                    const uint8_t* data = computeBindingOffset(accessor) + bufferData;
-                    const uint32_t size = computeBindingSize(accessor);
+                    auto* data = computeBindingOffset(accessor) + bufferData;
 
-                    // This creates a copy because we don't know when the user will free the cgltf
-                    // source data. For non-morphed vertex buffers, we use a sharing mechanism to
-                    // prevent copies, but here we just want to keep it as simple as possible.
-                    uint8_t* clone = (uint8_t*) malloc(size);
-                    memcpy(clone, data, size);
-
-                    BufferObject* bufferObject = BufferObject::Builder().size(size).build(engine);
-                    VertexBuffer::BufferDescriptor bd(clone, size, FREE_CALLBACK);
-                    bufferObject->setBuffer(engine, std::move(bd));
-                    morphHelperPrim.targets.push_back({bufferObject, targetIndex, atype});
+                    if (accessor->type == cgltf_type_vec3) {
+                        morphHelperPrim.targets->setPositionsAt(engine, targetIndex,
+                                (const float3*) data, vertexBuffer->getVertexCount());
+                    } else {
+                        assert_invariant(accessor->type == cgltf_type_vec4);
+                        morphHelperPrim.targets->setPositionsAt(engine, targetIndex,
+                                (const float4*) data, vertexBuffer->getVertexCount());
+                    }
                 }
             }
         }