+++ libs/utils/src/Log.cpp
@@ -46,6 +46,7 @@ private:
 };
 
 ostream& LogStream::flush() noexcept {
+    std::lock_guard lock(mLock);
     Buffer& buf = getBuffer();
 #ifdef __ANDROID__
     switch (mPriority) {
@@ -85,6 +86,14 @@ ostream& LogStream::flush() noexcept {
     buf.reset();
     return *this;
 }
+
+
+/*
+ * We can't use thread_local because on Android we're currently using several dynamic libraries
+ * including this .o (via libutils.a), which violates the ODR and ends-up with only one of
+ * the thread_local instance initialized.
+ * For this reason, ostream is protected by a mutex instead.
+ */
 
 static LogStream cout(LogStream::Priority::LOG_DEBUG);
 static LogStream cerr(LogStream::Priority::LOG_ERROR);